Remove deprecated model.format
